move find_files_like to files module
